feat: wire the playback presenter, and fix the bug writing its interface found (D63)
D62 built the presenter beside the form; this connects it and deletes what it
replaced, which is the order D59 requires — the old code goes in the commit that
proves the replacement covers it.

TransportBar implements IPlaybackView directly. It IS the view, so an adapter
would have been a second object doing nothing but renaming events.

  MainForm.cs            4,436 -> 4,379 lines
  playback in the form   3 handlers + AdvancePlayback + a Stopwatch -> one call
  tests over that logic  0 -> 16

WRITING THE INTERFACE FOUND A REAL DEFECT. IPlaybackView documents that setting
Playing must not raise PlayPauseToggled, or the presenter re-enters its own
handler. Writing that rule down forced a look at the control, and
TransportBar.Playing's setter raised the event — so the contract was violated by
the very control it was written for.

It conflated two different things: the user pressing the button, and somebody
assigning the property. SetDemoLength assigns it, and the presenter assigns it
when playback reaches an end.

The control already knew this distinction and had never applied it here: ShowTick
says in its own summary that it moves the readout "without raising Scrubbed".
Ticks had the rule; playing did not. Fixed with TogglePlayingByUser() for the
button, leaving the setter to update state and labels only.

That is the argument for interfaces, concretely. The bug was invisible while the
form and the control were one tangle, because the form never assigned Playing
from a path that could re-enter. It became visible the moment the boundary had to
be WRITTEN DOWN — the same mechanism as D54's compiler-enforced boundary one
level up: the compiler enforces what you write, and writing it is what makes you
look.

Speed is double throughout now rather than float: PlaybackClock.TimeScale and
TransportBar both use double, and the float in the first draft of the interface
was a lossy mismatch with both.

Verified: all seven suites green, plus the UI suite — 12 tests under
run-exclusive.ps1 — which is the one that actually drives the control whose event
contract changed, and which the ordinary suites cannot see.

Playback is one concern of about seven still in the form. The pattern is proven
end to end on the smallest of them, which was the point of doing it first.

+++ managed/Tf2DemoSalvage.Presentation/IPlaybackView.cs
@@ -25,10 +25,10 @@ public sealed class PlayingEventArgs(bool playing) : EventArgs
 
 /// <summary>The user changed the playback speed.</summary>
 /// <param name="speed">The multiplier, 1 being real time and negative running backwards.</param>
-public sealed class SpeedEventArgs(float speed) : EventArgs
+public sealed class SpeedEventArgs(double speed) : EventArgs
 {
     /// <summary>The speed multiplier.</summary>
-    public float Speed { get; } = speed;
+    public double Speed { get; } = speed;
 }
 
 /// <summary>The moment the scene should now show.</summary>
